requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
django-rest-framework==0.1.0
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pip==23.0.1
pluggy==1.0.0
pytest-django==4.5.2
pytest==7.3.1
pytz==2023.3
requests==2.29.0
setuptools==67.6.1
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
wheel==0.40.0
